+++ src/App.js
@@ -4,7 +4,7 @@ import AppRouter from "./Router";
 function App() {
   // const isUser = false;
 
-  let isUser = true;
+  let isUser = false;
   return <AppRouter isLoggedIn={Boolean(isUser)} />;
 }
