+++ src/components/Certification/CertificationForm.jsx
@@ -2,28 +2,26 @@ import { useEffect } from "react";
 import { useRef } from "react";
 import { useState } from "react";
 import { useHistory } from "react-router-dom";
+import useCertification from "../../Hooks/useCertification";
+import useSignup from "../../Hooks/useSignup";
 import "./CertificationForm.css";
 
 const CertificationForm = () => {
-  const [number, serNumber] = useState("");
-  const [minute, setMinute] = useState(0);
-  const [second, setSecond] = useState(0);
-  const [alert, setAlert] = useState("");
-  const [isPossibleReSend, setIsPossibleSend] = useState(true);
-  const time = useRef(299);
-  const timerId = useRef(null);
-  const reSendId = useRef();
-  const reSendCount = useRef(5);
-  const history = useHistory();
-
-  const onChange = (event) => {
-    const {
-      target: { value, name },
-    } = event;
-    if (name === "certification") {
-      serNumber(value);
-    }
-  };
+  const {
+    onChange,
+    number,
+    minute,
+    second,
+    time,
+    timerId,
+    reSendId,
+    alert,
+    setAlert,
+    setMinute,
+    setSecond,
+    onSubmit,
+    reSend,
+  } = useCertification();
 
   useEffect(() => {
     timerId.current = setInterval(() => {
@@ -40,38 +38,13 @@ const CertificationForm = () => {
 
   useEffect(() => {
     if (time.current <= 0) {
-      isPossibleReSend &&
-        window.alert(
-          "인증시간을 초과했습니다. '재전송'을 눌러 다시 인증해주시기 바랍니다."
-        );
+      window.alert(
+        "인증시간을 초과했습니다. '재전송'을 눌러 다시 인증해주시기 바랍니다."
+      );
       clearInterval(timerId.current);
       setAlert("시간을 초과했습니다. 다시 인증해주세요.");
     }
   }, [second]);
-
-  const reSend = () => {
-    if (isPossibleReSend) {
-      time.current = 299;
-      setAlert(
-        `인증번호가 발송 되었습니다. 가입하신 메일을 확인해주세요 (재전송 기회 : ${reSendCount.current}회)`
-      );
-      reSendCount.current -= 1;
-      if (reSendCount.current <= 0) {
-        setIsPossibleSend(false);
-      }
-    } else if (!isPossibleReSend) {
-      reSendId.current.disabled = true;
-      window.alert("재전송 기회 5번을 초과하셨습니다. 내일 다시 시도해주세요.");
-      setAlert("재전송 기회 5번을 초과하셨습니다. 내일 다시 시도해주세요.");
-    }
-  };
-
-  const onSubmit = (event) => {
-    event.preventDefault();
-    console.log("i submit");
-    history.push("/");
-  };
-
   return (
     <form id="certification-form" onSubmit={onSubmit}>
       <h1 id="certificationForm-title">인증</h1>
